sources: apply_network_config should optionally happen on every boot
This change makes it possible to reconfigure networking on every boot,
as was the behavior prior to the fix for LP#1571004.  Unlike the
original behavior, the default behavior is to reconfigure networking
PER_INSTANCE.  If 'maintain_network' is set to True in the datasource's
metadata, networking is reconfigured PER_ALWAYS.

This change also updates DataSourceSmartOS to retrieve the value of
maintain_network from sdc:maintain_network.

LP: #1765801
Joyent: OS-6902 cloud-init should be able to reconfigure network on each boot